Change default shell for execution
It seems, for those who don't have bash as their default shell, [[ is not valid syntax for `sh` (it did not run on my machine running Linux Mint 19 and zsh as my default, I got multiple `./docker-build.sh: 4: ./docker-build.sh: [[: not found` errors). 

I would propose changing this to `#!/bin/bash`, for which this is definitely valid syntax.