Micronaut sample (#303)
* First import of micronaut native demo

* Cleanup, upgrade dependencies and use latest GraalVM version

* Change default shell for execution 

It seems, for those who don't have bash as their default shell, [[ is not valid syntax for `sh` (it did not run on my machine running Linux Mint 19 and zsh as my default, I got multiple `./docker-build.sh: 4: ./docker-build.sh: [[: not found` errors). 

I would propose changing this to `#!/bin/bash`, for which this is definitely valid syntax.

* Further differentiates running via SAM vs AWS

It wasn't initially super obvious to me how to run this locally, but I was able to figure it out. I'm hoping to make it clearer to anyone else looking at it in the future via this change.

* Updated Micronaut readmes